backend/requirements.txt
fastapi
uvicorn
pydantic
requests
beautifulsoup4
python-dotenv
google-generativeai
numpy
